// BOnlineStore.UI/ClientApp/src/app/base-classes/base-enums/datasource-functions.enum.ts
export enum DatasourceFunctionsEnum {
  GETLOAD = 'GetLoad',
  NOLOAD = 'NoLoad',
  GETLOADPOST = 'GetLoadPost',
  NOLOADPOST = 'NoLoadPost',
  GETINSERT = 'GetInsert',
  NOINSERT = 'NoInsert',
  GETUPDATE = 'GetUpdate',
  NOUPDATE = 'NoUpdate',
  GETREMOVE = 'GetRemove',
  NOREMOVE = 'NoRemove',
}

export enum TableDefinitionsDataSourceTypeEnum {
  DataSource = 'DataSource',
  DataSourceRawMode = 'DataSourceRawMode',
  CustomDataSource = 'CustomDataSource',
  ArrayDataSource = 'ArrayDataSource',
}